After profiling, identified bottleneck and changed structure. Not yet ported to examples, but this should be the new way how things are done.

+++ examples/harmonic_oscillator_1d/bare_interaction.jl
@@ -11,7 +11,7 @@ function construct_bare_interaction(ho_orbital::HOOrbital1D, n_basis::Integer, c
     """
     W = fill(0.0, (2*n_basis, 2*n_basis))
     @nloops 2 i W begin
-        (@nref 2 W i) = coupling * ho_orbital(i_1, 0.0)*ho_orbital(i_2, 0.0) / sqrt(2)
+        (@nref 2 W i) = coupling * ho_orbital(Int16(i_1), 0.0)*ho_orbital(Int16(i_2), 0.0) / sqrt(2)
     end
     return W
 end
